+++ caoslib/main.py
@@ -4,12 +4,13 @@ from clint.textui import puts, colored
 #this is far from ideal but a little more abstraction
 if __name__ == '__main__':
     mode = arguments.Args()[0]
+    args = arguments.Args()
     flags = arguments.Args().flags
 
     #project init
     import initialization
     if mode in initialization.options:
-        initialization.handler(mode, flags)
+        initialization.handler(mode)
 
     #edjudje is needed
     import src
@@ -19,6 +20,6 @@ if __name__ == '__main__':
     #just working with material on computer
     import tasks
     if mode in tasks.options:
-        tasks.handler(mode, flags)
+        tasks.handler(mode, args)
 
     puts(colored.red(f"No valid arguments are found"))
